refactor: reorganize project structure by moving database and email services to infrastructure; update imports accordingly (#31)
