+++ src/services/authService.js
@@ -2,9 +2,9 @@ import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 import User from "../models/User.js";
 import TempCode from "../models/TempCode.js";
-import { generateNickname } from "../utils/nicknameGen.js";
-import { verifyGoogleToken } from "../utils/googleClient.js";
-import { sendVerificationEmail } from "../utils/emailService.js";
+import { generateNickname } from "../shared/utils/nicknameGen.js";
+import { verifyGoogleToken } from "../infrastructure/google/googleClient.js";
+import { sendVerificationEmail } from "../infrastructure/email/emailService.js";
 
 // Registration logic
 export const register = async (request, reply) => {
