+++ src/shared/utils/nicknameGen.js
